Estimate plasmid conjugation rates

estimate_conj_rate returns the conjugation rate estimates from data.

estimate_conj_rate(
  data,
  method = c("SM", "ASM"),
  id_cols = c("ID"),
  verbose = T
)

Arguments

data
Dataframe. Required columns vary depending on the method used to estimate the conjugation rates. The following columns are typically required: D.t, R.t, T.t are final population sizes; D.0, R.0 are initial pop sizes; psi.D, psi.R, psi.T are growth rates; t is the time of measurement (assumes start at 0). Optional columns: T.0, psi.max (for the Simonsen method).

method
String or list of strings. Describes method(s) used to estimate the conjugation rate. Takes the following values: - SM (Simonsen Method): The population growth rate can be supplied with psi.N or psi.max; otherwise the method will choose the maximum among psi.D, psi.R, psi.T; - ASM (Approximate Simonsen Method); - TD \(T/D\); - T_DR \(T/DR\); - T_RT \(T/(R+T)\): The fraction of plasmid-carrying recipients; - Dionisio; - Gama.

id_cols
List of strings. Column names in the data that should be treated as identifiers (will be returned in output).

verbose
Boolean. Should the method return warnings?

Value

Dataframe with the conjugation rate estimates. Contains 3 columns: sample IDs, estimates, methods.

See also

Examples

estimate_conj_rate(DRT_example)

#>   ID     estimate method
#> 1 A1 9.910688e-10     SM
#> 2 A2 9.910688e-10     SM
#> 3 A3 9.910688e-10     SM
#> 4 A1 9.004904e-10    ASM
#> 5 A2 9.004904e-10    ASM
#> 6 A3 8.008783e-10    ASM

estimate_conj_rate(DRT_example, method = c('SM', 'T_DR', 'Dionisio'))

#>   ID     estimate   method
#> 1 A1 9.910688e-10       SM
#> 2 A2 9.910688e-10       SM
#> 3 A3 9.910688e-10       SM
#> 4 A1 1.000000e-09     T_DR
#> 5 A2 1.000000e-09     T_DR
#> 6 A3 1.000000e-09     T_DR
#> 7 A1 1.000000e-02 Dionisio
#> 8 A2 1.000000e-02 Dionisio
#> 9 A3 1.000000e-02 Dionisio