Estimate critical times

estimate_crit_time returns the conjugation rates and critical times corresponding to a conjugation experiment. For the ASM to return a proper conjugation estimate the time point of measurement must be smaller than min(tcrit1, tcrit2, tcrit3).

estimate_crit_time(
  DRT,
  TRT = NULL,
  tol_factor = 10,
  id_cols = c("ID"),
  verbose = T
)

Arguments

DRT
Dataframe. Data corresponding to the first conjugation experiment. Necessary columns: D.0, R.0 are initial pop sizes; psi.D, psi.R, psi.T are growth rates. If gamma.D is not among the columns, the function will attempt to estimate gamma.D from the data given. This requires D.t, R.t, T.t (final population sizes) and t (the time of measurement). If gamma.T is among the columns, the input under TRT will be ignored.

TRT
Dataframe. Data corresponding to the second conjugation experiment. Same columns needed as for DRT.

tol_factor
Double. The imprecision factor: this factor indicates how close the system is allowed to approach a state where it violates the ASM assumptions (smaller values allow greater violation).

id_cols
List of strings. Specifies the column names that should be treated as identifiers (will be returned in output).

verbose
Boolean. Should warnings be returned?

Value

Dataframe of critical time estimates.

See also

Examples

estimate_crit_time(DRT_example, TRT_example)

#>   ID      gamma.D      gamma.T tcrit1 tcrit2 tcrit3 min_tcrit
#> 1 A1 9.004904e-10 1.000545e-09   9.80   11.6   11.5      9.80
#> 2 A2 9.004904e-10 1.000545e-08   7.49   11.6   10.2      7.49
#> 3 A3 8.008783e-10 1.001098e-08   7.49   11.8   10.2      7.49

estimate_crit_time(DRT_example, TRT_example, id_cols = c('ID', 't'))

#>   ID t      gamma.D      gamma.T tcrit1 tcrit2 tcrit3 min_tcrit
#> 1 A1 7 9.004904e-10 1.000545e-09   9.80   11.6   11.5      9.80
#> 2 A2 7 9.004904e-10 1.000545e-08   7.49   11.6   10.2      7.49
#> 3 A3 7 8.008783e-10 1.001098e-08   7.49   11.8   10.2      7.49