Growth fit for plotting get_growth_fit_for_plot computes a vector of the fit of the growth data, for comparison against the OD data.

Growth fit for plotting get_growth_fit_for_plot computes a vector of the fit of the growth data, for comparison against the OD data.

get_growth_fit_for_plot(
  OD_data,
  t_col = "time",
  growth_fit,
  blank_tf = T,
  verbose = T
)

Arguments

OD_data
Dataframe of OD data. Expects a column for the time (t_col), and will interpret all other columns as OD data for different samples (identified by their column name).

t_col
Name of the time column (defaults to 'time').

growth_fit
Output of estimate_growth_rate for the OD_data; requires columns name, carrying_cap, growth_rate, N0, t_min, t_stat

blank_tf
Boolean. Should a blank value be subtracted?

verbose
Boolean. Should warnings be returned?

Value

Simulated log growth matching the input fits.

See also

Examples

growth_fit <- estimate_growth_rate(growth_example, t_start = 1, t_col = 'Time_h')
get_growth_fit_for_plot(growth_example, t_col = 'Time_h', growth_fit)

#>              ID       time          OD
#> 1      Donor_OD  0.0000000 0.007447194
#> 2      Donor_OD  0.3243243 0.007447194
#> 3      Donor_OD  0.6486486 0.007447194
#> 4      Donor_OD  0.9729730 0.007447194
#> 5      Donor_OD  1.2972973 0.089975804
#> 6      Donor_OD  1.6216216 0.099680837
#> 7      Donor_OD  1.9459459 0.109964026
#> 8      Donor_OD  2.2702703 0.120737926
#> 9      Donor_OD  2.5945946 0.131893726
#> 10     Donor_OD  2.9189189 0.143304912
#> 11     Donor_OD  3.2432432 0.154832612
#> 12     Donor_OD  3.5675676 0.166332174
#> 13     Donor_OD  3.8918919 0.177660352
#> 14     Donor_OD  4.2162162 0.188682335
#> 15     Donor_OD  4.5405405 0.199277920
#> 16     Donor_OD  4.8648649 0.209346238
#> 17     Donor_OD  5.1891892 0.218808714
#> 18     Donor_OD  5.5135135 0.227610180
#> 19     Donor_OD  5.8378378 0.235718299
#> 20     Donor_OD  6.1621622 0.243121652
#> 21     Donor_OD  6.4864865 0.249826886
#> 22     Donor_OD  6.8108108 0.255855360
#> 23     Donor_OD  7.1351351 0.261239669
#> 24     Donor_OD  7.4594595 0.266020328
#> 25     Donor_OD  7.7837838 0.270242804
#> 26     Donor_OD  8.1081081 0.273955026
#> 27     Donor_OD  8.4324324 0.277205379
#> 28     Donor_OD  8.7567568 0.280041188
#> 29     Donor_OD  9.0810811 0.282507629
#> 30     Donor_OD  9.4054054 0.284647006
#> 31     Donor_OD  9.7297297 0.286498334
#> 32     Donor_OD 10.0540540 0.288097138
#> 33     Donor_OD 10.3783784 0.289475440
#> 34     Donor_OD 10.7027027 0.290661851
#> 35     Donor_OD 11.0270270 0.291681756
#> 36     Donor_OD 11.3513513 0.292557540
#> 37     Donor_OD 11.6756757 0.293308845
#> 38     Donor_OD 12.0000000 0.293952831
#> 39 Recipient_OD  0.0000000 0.007447194
#> 40 Recipient_OD  0.3243243 0.007447194
#> 41 Recipient_OD  0.6486486 0.007447194
#> 42 Recipient_OD  0.9729730 0.007447194
#> 43 Recipient_OD  1.2972973 0.045261797
#> 44 Recipient_OD  1.6216216 0.063389175
#> 45 Recipient_OD  1.9459459 0.084765392
#> 46 Recipient_OD  2.2702703 0.106621052
#> 47 Recipient_OD  2.5945946 0.125938110
#> 48 Recipient_OD  2.9189189 0.140933361
#> 49 Recipient_OD  3.2432432 0.151443115
#> 50 Recipient_OD  3.5675676 0.158292669
#> 51 Recipient_OD  3.8918919 0.162547663
#> 52 Recipient_OD  4.2162162 0.165112601
#> 53 Recipient_OD  4.5405405 0.166630824
#> 54 Recipient_OD  4.8648649 0.167519800
#> 55 Recipient_OD  5.1891892 0.168037030
#> 56 Recipient_OD  5.5135135 0.168336855
#> 57 Recipient_OD  5.8378378 0.168510284
#> 58 Recipient_OD  6.1621622 0.168610475
#> 59 Recipient_OD  6.4864865 0.168668316
#> 60 Recipient_OD  6.8108108 0.168701693
#> 61 Recipient_OD  7.1351351 0.168720949
#> 62 Recipient_OD  7.4594595 0.168732057
#> 63 Recipient_OD  7.7837838 0.168738463
#> 64 Recipient_OD  8.1081081 0.168742159
#> 65 Recipient_OD  8.4324324 0.168744290
#> 66 Recipient_OD  8.7567568 0.168745519
#> 67 Recipient_OD  9.0810811 0.168746228
#> 68 Recipient_OD  9.4054054 0.168746637
#> 69 Recipient_OD  9.7297297 0.168746873
#> 70 Recipient_OD 10.0540540 0.168747009
#> 71 Recipient_OD 10.3783784 0.168747087
#> 72 Recipient_OD 10.7027027 0.168747132
#> 73 Recipient_OD 11.0270270 0.168747158
#> 74 Recipient_OD 11.3513513 0.168747173
#> 75 Recipient_OD 11.6756757 0.168747182
#> 76 Recipient_OD 12.0000000 0.168747187