Scan critical times

scan_crit_time returns the conjugation rates and critical times corresponding to a DRT conjugation experiment, assuming different gamma.T values. For the ASM to return a proper conjugation estimate the time point of measurement must be smaller than min(tcrit1, tcrit2, tcrit3).

scan_crit_time(
  DRT,
  tol_factor = 10,
  id_cols = c("ID"),
  mult_seq = 10^seq(-2, 5, 1)
)

Arguments

DRT
Dataframe. Data corresponding to the first conjugation experiment. Necessary columns: D.0, R.0 are initial pop sizes; psi.D, psi.R, psi.T are growth rates. If gamma.D is not among the columns, the function will attempt to estimate gamma.D from the data given. This requires D.t, R.t, T.t (final population sizes) and t (the time of measurement). If gamma.T is among the columns, the input under TRT will be ignored.

tol_factor
Double. The imprecision factor: this factor indicates how close the system is allowed to approach a state where it violates the ASM assumptions (smaller values allow greater violation).

id_cols
List of strings. Specifies the column names that should be treated as identifiers (will be returned in output).

mult_seq
Vector. Describes the multiplication factors to take into account for gamma.T.

Value

Dataframe of estimated critical times for the input DRT data and a range of gamma.T values.

See also

Examples

scan_crit_time(DRT_example)

#>    ID mult_factor      gamma.D      gamma.T tcrit1 tcrit2 tcrit3 min_tcrit
#> 1  A1       1e-02 9.004904e-10 9.004904e-12 14.500   11.6  14.10    11.600
#> 2  A1       1e-01 9.004904e-10 9.004904e-11 12.200   11.6  12.90    11.600
#> 3  A1       1e+00 9.004904e-10 9.004904e-10  9.900   11.6  11.60     9.900
#> 4  A1       1e+01 9.004904e-10 9.004904e-09  7.600   11.6  10.30     7.600
#> 5  A1       1e+02 9.004904e-10 9.004904e-08  5.300   11.6   9.01     5.300
#> 6  A1       1e+03 9.004904e-10 9.004904e-07  3.000   11.6   7.73     3.000
#> 7  A1       1e+04 9.004904e-10 9.004904e-06  0.693   11.6   6.45     0.693
#> 8  A1       1e+05 9.004904e-10 9.004904e-05 -1.610   11.6   5.17    -1.610
#> 9  A2       1e-02 9.004904e-10 9.004904e-12 14.500   11.6  14.10    11.600
#> 10 A2       1e-01 9.004904e-10 9.004904e-11 12.200   11.6  12.90    11.600
#> 11 A2       1e+00 9.004904e-10 9.004904e-10  9.900   11.6  11.60     9.900
#> 12 A2       1e+01 9.004904e-10 9.004904e-09  7.600   11.6  10.30     7.600
#> 13 A2       1e+02 9.004904e-10 9.004904e-08  5.300   11.6   9.01     5.300
#> 14 A2       1e+03 9.004904e-10 9.004904e-07  3.000   11.6   7.73     3.000
#> 15 A2       1e+04 9.004904e-10 9.004904e-06  0.693   11.6   6.45     0.693
#> 16 A2       1e+05 9.004904e-10 9.004904e-05 -1.610   11.6   5.17    -1.610
#> 17 A3       1e-02 8.008783e-10 8.008783e-12 14.600   11.8  14.20    11.800
#> 18 A3       1e-01 8.008783e-10 8.008783e-11 12.300   11.8  12.90    11.800
#> 19 A3       1e+00 8.008783e-10 8.008783e-10 10.000   11.8  11.60    10.000
#> 20 A3       1e+01 8.008783e-10 8.008783e-09  7.720   11.8  10.40     7.720
#> 21 A3       1e+02 8.008783e-10 8.008783e-08  5.420   11.8   9.08     5.420
#> 22 A3       1e+03 8.008783e-10 8.008783e-07  3.110   11.8   7.80     3.110
#> 23 A3       1e+04 8.008783e-10 8.008783e-06  0.810   11.8   6.52     0.810
#> 24 A3       1e+05 8.008783e-10 8.008783e-05 -1.490   11.8   5.24    -1.490

scan_crit_time(DRT_example, id_cols = c('ID', 't'), mult_seq = 10**seq(0, 3, 0.1))

#>    ID t mult_factor      gamma.D      gamma.T tcrit1 tcrit2 tcrit3 min_tcrit
#> 1  A1 7    1.000000 9.004904e-10 9.004904e-10   9.90   11.6  11.60      9.90
#> 2  A1 7    1.258925 9.004904e-10 1.133650e-09   9.67   11.6  11.40      9.67
#> 3  A1 7    1.584893 9.004904e-10 1.427181e-09   9.44   11.6  11.30      9.44
#> 4  A1 7    1.995262 9.004904e-10 1.796715e-09   9.21   11.6  11.20      9.21
#> 5  A1 7   10.000000 9.004904e-10 9.004904e-09   7.60   11.6  10.30      7.60
#> 6  A1 7  100.000000 9.004904e-10 9.004904e-08   5.30   11.6   9.01      5.30
#> 7  A1 7 1000.000000 9.004904e-10 9.004904e-07   3.00   11.6   7.73      3.00
#> 8  A1 7   12.589254 9.004904e-10 1.133650e-08   7.37   11.6  10.20      7.37
#> 9  A1 7  125.892541 9.004904e-10 1.133650e-07   5.07   11.6   8.88      5.07
#> 10 A1 7   15.848932 9.004904e-10 1.427181e-08   7.14   11.6  10.00      7.14
#> 11 A1 7  158.489319 9.004904e-10 1.427181e-07   4.84   11.6   8.76      4.84
#> 12 A1 7   19.952623 9.004904e-10 1.796715e-08   6.91   11.6   9.91      6.91
#> 13 A1 7  199.526231 9.004904e-10 1.796715e-07   4.61   11.6   8.63      4.61
#> 14 A1 7    2.511886 9.004904e-10 2.261930e-09   8.98   11.6  11.10      8.98
#> 15 A1 7   25.118864 9.004904e-10 2.261930e-08   6.68   11.6   9.78      6.68
#> 16 A1 7  251.188643 9.004904e-10 2.261930e-07   4.38   11.6   8.50      4.38
#> 17 A1 7    3.162278 9.004904e-10 2.847601e-09   8.75   11.6  10.90      8.75
#> 18 A1 7    3.981072 9.004904e-10 3.584917e-09   8.52   11.6  10.80      8.52
#> 19 A1 7   31.622777 9.004904e-10 2.847601e-08   6.45   11.6   9.65      6.45
#> 20 A1 7  316.227766 9.004904e-10 2.847601e-07   4.15   11.6   8.37      4.15
#> 21 A1 7   39.810717 9.004904e-10 3.584917e-08   6.22   11.6   9.52      6.22
#> 22 A1 7  398.107171 9.004904e-10 3.584917e-07   3.92   11.6   8.24      3.92
#> 23 A1 7    5.011872 9.004904e-10 4.513143e-09   8.29   11.6  10.70      8.29
#> 24 A1 7   50.118723 9.004904e-10 4.513143e-08   5.99   11.6   9.40      5.99
#> 25 A1 7  501.187234 9.004904e-10 4.513143e-07   3.69   11.6   8.12      3.69
#> 26 A1 7    6.309573 9.004904e-10 5.681710e-09   8.06   11.6  10.50      8.06
#> 27 A1 7   63.095734 9.004904e-10 5.681710e-08   5.76   11.6   9.27      5.76
#> 28 A1 7  630.957344 9.004904e-10 5.681710e-07   3.46   11.6   7.99      3.46
#> 29 A1 7    7.943282 9.004904e-10 7.152849e-09   7.83   11.6  10.40      7.83
#> 30 A1 7   79.432823 9.004904e-10 7.152849e-08   5.53   11.6   9.14      5.53
#> 31 A1 7  794.328235 9.004904e-10 7.152849e-07   3.23   11.6   7.86      3.23
#> 32 A2 7    1.000000 9.004904e-10 9.004904e-10   9.90   11.6  11.60      9.90
#> 33 A2 7    1.258925 9.004904e-10 1.133650e-09   9.67   11.6  11.40      9.67
#> 34 A2 7    1.584893 9.004904e-10 1.427181e-09   9.44   11.6  11.30      9.44
#> 35 A2 7    1.995262 9.004904e-10 1.796715e-09   9.21   11.6  11.20      9.21
#> 36 A2 7   10.000000 9.004904e-10 9.004904e-09   7.60   11.6  10.30      7.60
#> 37 A2 7  100.000000 9.004904e-10 9.004904e-08   5.30   11.6   9.01      5.30
#> 38 A2 7 1000.000000 9.004904e-10 9.004904e-07   3.00   11.6   7.73      3.00
#> 39 A2 7   12.589254 9.004904e-10 1.133650e-08   7.37   11.6  10.20      7.37
#> 40 A2 7  125.892541 9.004904e-10 1.133650e-07   5.07   11.6   8.88      5.07
#> 41 A2 7   15.848932 9.004904e-10 1.427181e-08   7.14   11.6  10.00      7.14
#> 42 A2 7  158.489319 9.004904e-10 1.427181e-07   4.84   11.6   8.76      4.84
#> 43 A2 7   19.952623 9.004904e-10 1.796715e-08   6.91   11.6   9.91      6.91
#> 44 A2 7  199.526231 9.004904e-10 1.796715e-07   4.61   11.6   8.63      4.61
#> 45 A2 7    2.511886 9.004904e-10 2.261930e-09   8.98   11.6  11.10      8.98
#> 46 A2 7   25.118864 9.004904e-10 2.261930e-08   6.68   11.6   9.78      6.68
#> 47 A2 7  251.188643 9.004904e-10 2.261930e-07   4.38   11.6   8.50      4.38
#> 48 A2 7    3.162278 9.004904e-10 2.847601e-09   8.75   11.6  10.90      8.75
#> 49 A2 7    3.981072 9.004904e-10 3.584917e-09   8.52   11.6  10.80      8.52
#> 50 A2 7   31.622777 9.004904e-10 2.847601e-08   6.45   11.6   9.65      6.45
#> 51 A2 7  316.227766 9.004904e-10 2.847601e-07   4.15   11.6   8.37      4.15
#> 52 A2 7   39.810717 9.004904e-10 3.584917e-08   6.22   11.6   9.52      6.22
#> 53 A2 7  398.107171 9.004904e-10 3.584917e-07   3.92   11.6   8.24      3.92
#> 54 A2 7    5.011872 9.004904e-10 4.513143e-09   8.29   11.6  10.70      8.29
#> 55 A2 7   50.118723 9.004904e-10 4.513143e-08   5.99   11.6   9.40      5.99
#> 56 A2 7  501.187234 9.004904e-10 4.513143e-07   3.69   11.6   8.12      3.69
#> 57 A2 7    6.309573 9.004904e-10 5.681710e-09   8.06   11.6  10.50      8.06
#> 58 A2 7   63.095734 9.004904e-10 5.681710e-08   5.76   11.6   9.27      5.76
#> 59 A2 7  630.957344 9.004904e-10 5.681710e-07   3.46   11.6   7.99      3.46
#> 60 A2 7    7.943282 9.004904e-10 7.152849e-09   7.83   11.6  10.40      7.83
#> 61 A2 7   79.432823 9.004904e-10 7.152849e-08   5.53   11.6   9.14      5.53
#> 62 A2 7  794.328235 9.004904e-10 7.152849e-07   3.23   11.6   7.86      3.23
#> 63 A3 7    1.000000 8.008783e-10 8.008783e-10  10.00   11.8  11.60     10.00
#> 64 A3 7    1.258925 8.008783e-10 1.008246e-09   9.79   11.8  11.50      9.79
#> 65 A3 7    1.584893 8.008783e-10 1.269307e-09   9.56   11.8  11.40      9.56
#> 66 A3 7    1.995262 8.008783e-10 1.597962e-09   9.33   11.8  11.30      9.33
#> 67 A3 7   10.000000 8.008783e-10 8.008783e-09   7.72   11.8  10.40      7.72
#> 68 A3 7  100.000000 8.008783e-10 8.008783e-08   5.42   11.8   9.08      5.42
#> 69 A3 7 1000.000000 8.008783e-10 8.008783e-07   3.11   11.8   7.80      3.11
#> 70 A3 7   12.589254 8.008783e-10 1.008246e-08   7.49   11.8  10.20      7.49
#> 71 A3 7  125.892541 8.008783e-10 1.008246e-07   5.18   11.8   8.95      5.18
#> 72 A3 7   15.848932 8.008783e-10 1.269307e-08   7.26   11.8  10.10      7.26
#> 73 A3 7  158.489319 8.008783e-10 1.269307e-07   4.95   11.8   8.82      4.95
#> 74 A3 7   19.952623 8.008783e-10 1.597962e-08   7.03   11.8   9.97      7.03
#> 75 A3 7  199.526231 8.008783e-10 1.597962e-07   4.72   11.8   8.69      4.72
#> 76 A3 7    2.511886 8.008783e-10 2.011715e-09   9.10   11.8  11.10      9.10
#> 77 A3 7   25.118864 8.008783e-10 2.011715e-08   6.80   11.8   9.84      6.80
#> 78 A3 7  251.188643 8.008783e-10 2.011715e-07   4.49   11.8   8.57      4.49
#> 79 A3 7    3.162278 8.008783e-10 2.532599e-09   8.87   11.8  11.00      8.87
#> 80 A3 7    3.981072 8.008783e-10 3.188354e-09   8.64   11.8  10.90      8.64
#> 81 A3 7   31.622777 8.008783e-10 2.532599e-08   6.57   11.8   9.72      6.57
#> 82 A3 7  316.227766 8.008783e-10 2.532599e-07   4.26   11.8   8.44      4.26
#> 83 A3 7   39.810717 8.008783e-10 3.188354e-08   6.34   11.8   9.59      6.34
#> 84 A3 7  398.107171 8.008783e-10 3.188354e-07   4.03   11.8   8.31      4.03
#> 85 A3 7    5.011872 8.008783e-10 4.013900e-09   8.41   11.8  10.70      8.41
#> 86 A3 7   50.118723 8.008783e-10 4.013900e-08   6.11   11.8   9.46      6.11
#> 87 A3 7  501.187234 8.008783e-10 4.013900e-07   3.80   11.8   8.18      3.80
#> 88 A3 7    6.309573 8.008783e-10 5.053200e-09   8.18   11.8  10.60      8.18
#> 89 A3 7   63.095734 8.008783e-10 5.053200e-08   5.88   11.8   9.33      5.88
#> 90 A3 7  630.957344 8.008783e-10 5.053200e-07   3.57   11.8   8.05      3.57
#> 91 A3 7    7.943282 8.008783e-10 6.361602e-09   7.95   11.8  10.50      7.95
#> 92 A3 7   79.432823 8.008783e-10 6.361602e-08   5.65   11.8   9.21      5.65
#> 93 A3 7  794.328235 8.008783e-10 6.361602e-07   3.34   11.8   7.93      3.34